Refactor project into modules and add tests
