+++ README.md
@@ -1,0 +1,27 @@
+# Audiobook QC
+
+This project provides a Tkinter application to align audiobook scripts with ASR transcripts.
+
+## Requirements
+
+- Python 3.12
+- `unidecode`
+- `pdfplumber`
+- `rapidfuzz`
+- `flake8` and `black` for linting (optional)
+
+Install dependencies with:
+
+```bash
+pip install -r requirements.txt
+```
+
+## Usage
+
+Run the GUI with:
+
+```bash
+python qc_app.py
+```
+
+The application lets you select a script (PDF or TXT) and an ASR transcript, performs alignment and saves a `.qc.json` file.
