+++ qc_app.py
@@ -1,0 +1,6 @@
+"""Entry point for the QC Audiobook application."""
+
+from gui import App
+
+if __name__ == "__main__":
+    App().mainloop()
