+++ requirements.txt
@@ -1,0 +1,3 @@
+unidecode
+pdfplumber
+rapidfuzz
